server/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.0.14
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-taggit==5.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
Markdown==3.6
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
